DTO & Mappers added

+++ ExpertSystem/src/main/java/ru/margarita/ExpertSystem/repository/DivisionRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.margarita.ExpertSystem.domain.Division;
 
 @Repository
-public interface DivisionRepo extends CrudRepository<Division, Long> {
+public interface DivisionRepo extends CrudRepository<Division, Integer> {
 }
