+++ ExpertSystem/src/main/java/ru/margarita/ExpertSystem/repository/MessageRepo.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.margarita.ExpertSystem.domain.Message;
 
 @Repository
-public interface MessageRepo extends CrudRepository<Message, Long> {
+public interface MessageRepo extends CrudRepository<Message, Integer> {
 }
